remove dependency to cscore when testing finder
